fix(tests): return correct exit status from busted
We use `busted.subscribe` to override the output handlers with a
callback. To implement the mediator pattern, Busted uses
[mediator_lua](https://github.com/Olivine-Labs/mediator_lua).

The second value returned by the subscription callback is used to decide
whether to continue execution of other subscribers. Since we only return
`nil`, the test failure was not handled to exit with the right status
and failing tests were exiting with `0`.

This commit changes the return value of the callback to: `nil, true` so
that the original callback is executed to handle the test result and
return the correct exit status.